Fix a circular import problem
